postCount in category + unique category + check category empty to delete + payload editPost

+++ controller/categoryController.js
@@ -1,16 +1,26 @@
 import schema from "../models/category.js";
 import crypto from "crypto";
-import { getCountByCategory } from "../helpers/index.js"
+import {
+    getProductCountByCategory,
+    getPostCountByCategory
+} from "../helpers/index.js"
 
-import { createItem, readItem, getAllItems, deleteItem } from "./services.js";
+import {
+    createItem,
+    readItem,
+    getAllItems,
+    deleteItem,
+    existCategory
+} from "./services.js";
 
 export const getCategories = async(req, res)  => {
     try {
         const categories = await getAllItems(schema);
         const payload = await Promise.all(categories.map(async(category) => {
             const { name, id } = category;
-            const productCount = await getCountByCategory(name);
-            return { name, id, productCount }
+            const productCount = await getProductCountByCategory(name);
+            const postCount = await getPostCountByCategory(name);
+            return { name, id, productCount, postCount };
         }))
         res.status(200).json(payload);
     } catch (error) { 
@@ -22,8 +32,14 @@ export const getCategories = async(req, res)  => {
 export const getCategoryById = async(req, res)  => {
     try {
         const { id } = req.params;
-        const result = await readItem(id, schema);
-        res.status(200).json(result);
+        const { name } = await readItem(id, schema);
+        if (name && id) {
+            const productCount = await getProductCountByCategory(name);
+            const postCount = await getPostCountByCategory(name);
+            const result = { name, id, postCount, productCount }
+            res.status(200).json(result);
+
+        }
     } catch (error) {
         console.log(error);
         res.status(500).json({ error });
@@ -33,12 +49,15 @@ export const getCategoryById = async(req, res)  => {
 export const addCategory = async(req, res) => {
     try {
         const { name } = req.body;
-        const id = crypto.createHash("md5")
+        if (!await existCategory(name)) {
+            const id = crypto.createHash("md5")
             .update(name)
             .digest("hex");
-        const payload = { id, name };
-        const result = await createItem(payload, schema);
-        res.status(200).json(result);
+            const payload = { id, name };
+            const result = await createItem(payload, schema);
+            return res.status(200).json(result);
+        }
+        throw "Already existing category";
     } catch (error) {
         console.log(error);
         res.status(500).json({ error });
@@ -49,9 +68,9 @@ export const deleteCategory = async(req, res) => {
     try {
         const { id } = req.params;
         const { name } = await readItem(id, schema);
-        const productCount = await getCountByCategory(name);
-        // da fare anche con i post!!!
-        if (!productCount) {
+        const productCount = await getProductCountByCategory(name);
+        const postCount = await getPostCountByCategory(name);
+        if (!productCount && !postCount) {
             const result = await deleteItem(id, schema);
             if (result) {
                 return res.status(200).json(result);
